Bugfix: only notify stalks created before the submission was sent
Previously we sent notifications looking at the chat's creation time.
This approach is wrong. If a chat started stalking a UVaUser and stopped for a while,
after resuming the stalk and a new submission taking place, the chat would be notified
of all of the submissions from the UVaUser during the period in which the chat had stopped stalking.