Updated irazasyed/telegram-bot-sdk to 2.0.x-dev
Note that even though the latest ^2.0 release of
irazasyed/telegram-bot-sdk at the time of this commit (v2.3,
https://github.com/irazasyed/telegram-bot-sdk/releases/tag/v2.3.0) is
marked as supporting Laravel 6.0, it does not: the Command class has a
reference to studly_case
(https://github.com/irazasyed/telegram-bot-sdk/blob/cb65c587fcbb6d74f519402fba429aeab411b4ae/src/Commands/Command.php#L182),
a string helper method that was removed in Laravel 5.8
(https://laravel.com/docs/5.8/upgrade#string-and-array-helpers).
The latest commit as of writing in ^2.0 (to which 2.0.x-dev points to)
does away with this problem.

Switching over to ^3.0 releases is an option I tried, but they are
unstable and undocumented. It also involves a minor rewrite of the bot
which I'm not comfortable undertaking now, given the state of those
branches and how poorly maintained the whole irazasyed/telegram-bot-sdk
project is.

Doing away with this dependency is probably for the better moving
forward, if the repository keeps being unmaintained.