Recreate the project with a clean installation of Laravel 6.0
In this commit, I deleted all of the files in the repository, did a
clean Laravel 6.0 installation doing `laravel new`, and pulled out all
of the application's code, while ensuring it works with the latest files
in the Laravel 6.0 tree.

This commit introduces many changes across many files. Most of the
modifications are cosmetic (i.e. import/method reorderings, renames,
comments). Many files that are no longer used are also deleted. A few
files are new.

Below, I've tried to summarise the most important changes to be aware
of:

* Remove files that no longer exist in the Laravel 6.0 tree
    * These files no longer come with a clean Laravel 6.0 project
    installation.
* Add .editorconfig
* Updated .env.example
* Updated .gitattributes and .gitignore
* Add .styleci.yml
* Change logging config
    * Laravel will store daily log files under storage/logs, with a 60
    day retention period.
* Update package.json and package-lock.json
* Exception handling: report all exceptions
* Laravel Tinker uses service auto-discovery
    * The laravel/tinker package uses service auto-discovery in its
    composer.json file. There is no need to register the service
    provider in config/app.php
* Update composer.json and composer.lock
* Misc (mostly cosmetic) changes to many files.